:hammer: Add scripts to dev and build the project
