+++ README.md
@@ -1,1 +1,19 @@
 # js13k-2020
+
+## Install
+
+```sh
+npm install
+```
+
+## Usage
+
+```sh
+npm run dev
+```
+
+## Production
+
+```js
+npm run build && npm run serve
+```
