--- index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <title>Pataplateforme</title>
-    </head>
-    <body>
-        <canvas>
-            <script type="application/javascript" src="main.js"></script>
-        </canvas>
-    </body>
-</html>
