+++ package.json
@@ -6,9 +6,16 @@
   "dependencies": {
     "kontra": "^7.1.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "parcel-bundler": "^1.12.4",
+    "prettier": "2.0.5",
+    "serve": "^11.3.2"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "parcel src/index.html",
+    "build": "parcel build src/index.html --no-source-maps",
+    "serve": "serve dist",
+    "format": "prettier --write src"
   },
   "repository": {
     "type": "git",
