+++ src/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8" />
+    <title>Pataplateforme</title>
+  </head>
+  <body>
+    <canvas>
+      <script src="./index.js"></script>
+    </canvas>
+  </body>
+</html>
